Add Model-M Mac layout and use home row for Emacs
Add a new layout for the IBM Model-M on macOS.

I finally decided how to standardize my layouts between PC and Mac.
The answer was to make the home row mods consistent for Emacs since I
use it so much and to use the either get used to those for the OS GUI
or to use the standard modifier keys next to the space bar for
navigating the GUI.

Later I will probably add some changes to handle driving SKHD and
Yabai better from the keyboard.